Tinkering with API.

+++ include/death/CovenThrone.h
@@ -39,7 +39,7 @@ namespace Death {
 
         void onStartHero(PlayerId player) override;
 
-        int getDamageTableIndex() const;
+        size_t getDamageTableIndex() const;
 
     private:
