+++ include/slavestodarkness/ChaosSorcererOnManticore.h
@@ -33,7 +33,7 @@ namespace SlavesToDarkness {
 
         void onRestore() override;
 
-        int getDamageTableIndex() const;
+        size_t getDamageTableIndex() const;
 
         Rerolls toHitRerolls(const Weapon *weapon, const Unit *target) const override;
