+++ include/slavestodarkness/Slaughterbrute.h
@@ -33,7 +33,7 @@ namespace SlavesToDarkness {
 
         void onRestore() override;
 
-        int getDamageTableIndex() const;
+        size_t getDamageTableIndex() const;
 
     private:
