+++ src/eldritchcouncil/ArchmageOnDragon.cpp
@@ -130,7 +130,7 @@ namespace EldritchCouncil {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ArchmageOnDragon::getDamageTableIndex() const {
+    size_t ArchmageOnDragon::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
