+++ src/gloomspitegitz/AleguzzlerGargant.cpp
@@ -89,7 +89,7 @@ namespace GloomspiteGitz {
         }
     }
 
-    int AleguzzlerGargant::getDamageTableIndex() const {
+    size_t AleguzzlerGargant::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
