+++ src/gloomspitegitz/ArachnarokSpiderWithFlinger.cpp
@@ -76,7 +76,7 @@ namespace GloomspiteGitz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ArachnarokSpiderWithFlinger::getDamageTableIndex() const {
+    size_t ArachnarokSpiderWithFlinger::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
