+++ src/gloomspitegitz/ColossalSquig.cpp
@@ -92,7 +92,7 @@ namespace GloomspiteGitz {
         return GloomspiteGitzBase::weaponDamage(weapon, target, hitRoll, woundRoll);
     }
 
-    int ColossalSquig::getDamageTableIndex() const {
+    size_t ColossalSquig::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
