+++ src/gloomspitegitz/LoonbossOnManglerSquigs.cpp
@@ -119,7 +119,7 @@ namespace GloomspiteGitz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int LoonbossOnManglerSquigs::getDamageTableIndex() const {
+    size_t LoonbossOnManglerSquigs::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
