+++ src/gloomspitegitz/ManglerSquigs.cpp
@@ -81,7 +81,7 @@ namespace GloomspiteGitz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ManglerSquigs::getDamageTableIndex() const {
+    size_t ManglerSquigs::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
