+++ src/gloomspitegitz/SkitterstrandArachnarok.cpp
@@ -65,7 +65,7 @@ namespace GloomspiteGitz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int SkitterstrandArachnarok::getDamageTableIndex() const {
+    size_t SkitterstrandArachnarok::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
