+++ src/gloomspitegitz/WebspinnerShamanOnSpider.cpp
@@ -94,7 +94,7 @@ namespace GloomspiteGitz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int WebspinnerShamanOnArachnarokSpider::getDamageTableIndex() const {
+    size_t WebspinnerShamanOnArachnarokSpider::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
