+++ src/greenskinz/OrrukWarbossWyvern.cpp
@@ -94,7 +94,7 @@ namespace Greenskinz {
         }
     }
 
-    int OrrukWarbossOnWyvern::getDamageTableIndex() const {
+    size_t OrrukWarbossOnWyvern::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
