+++ src/idonethdeepkin/AkhelianLeviadon.cpp
@@ -127,7 +127,7 @@ namespace IdonethDeepkin {
         onWounded();
     }
 
-    int AkhelianLeviadon::getDamageTableIndex() const {
+    size_t AkhelianLeviadon::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
