+++ src/ironjawz/Gordrakk.cpp
@@ -140,7 +140,7 @@ namespace Ironjawz {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int GordrakkTheFistOfGork::getDamageTableIndex() const {
+    size_t GordrakkTheFistOfGork::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
