+++ src/kharadron/ArkanautFrigate.cpp
@@ -145,7 +145,7 @@ namespace KharadronOverlords {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ArkanautFrigate::getDamageTableIndex() const {
+    size_t ArkanautFrigate::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
