+++ src/kharadron/ArkanautIronclad.cpp
@@ -152,7 +152,7 @@ namespace KharadronOverlords {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ArkanautIronclad::getDamageTableIndex() const {
+    size_t ArkanautIronclad::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
