+++ src/mawtribes/ThundertuskBeastriders.cpp
@@ -114,7 +114,7 @@ namespace OgorMawtribes {
         onWounded();
     }
 
-    int ThundertuskBeastriders::getDamageTableIndex() const {
+    size_t ThundertuskBeastriders::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
