+++ src/nighthaunt/BlackCoach.cpp
@@ -96,7 +96,7 @@ namespace Nighthaunt {
         m_move = g_damageTable[damageIndex].m_move;
     }
 
-    int BlackCoach::getDamageTableIndex() const {
+    size_t BlackCoach::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
