+++ src/slavestodarkness/ChaosLordOnManticore.cpp
@@ -148,7 +148,7 @@ namespace SlavesToDarkness {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ChaosLordOnManticore::getDamageTableIndex() const {
+    size_t ChaosLordOnManticore::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
