+++ src/slavestodarkness/ChaosSorcererOnManticore.cpp
@@ -169,7 +169,7 @@ namespace SlavesToDarkness {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ChaosSorcererOnManticore::getDamageTableIndex() const {
+    size_t ChaosSorcererOnManticore::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
