+++ src/slavestodarkness/ChaosWarshrine.cpp
@@ -106,7 +106,7 @@ namespace SlavesToDarkness {
         m_move = g_damageTable[getDamageTableIndex()].m_move;
     }
 
-    int ChaosWarshrine::getDamageTableIndex() const {
+    size_t ChaosWarshrine::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
