+++ src/slavestodarkness/SlaughterBrute.cpp
@@ -110,7 +110,7 @@ namespace SlavesToDarkness {
         onWounded();
     }
 
-    int Slaughterbrute::getDamageTableIndex() const {
+    size_t Slaughterbrute::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
