+++ src/slavestodarkness/VortexBeast.cpp
@@ -174,7 +174,7 @@ namespace SlavesToDarkness {
         onWounded();
     }
 
-    int MutalithVortexBeast::getDamageTableIndex() const {
+    size_t MutalithVortexBeast::getDamageTableIndex() const {
         auto woundsInflicted = wounds() - remainingWounds();
         for (auto i = 0u; i < g_numTableEntries; i++) {
             if (woundsInflicted < g_woundThresholds[i]) {
